Refactoring/cleanup of some view builders.
I have re-worked the "addresses", "flavors", "images", and "servers" view
builders to share logic which was previously duplicated across many ViewBuilder
classes. Also, I have re-worked the ViewBuilder class in general to persist
longer (meaning that a new ViewBuilder is not created for every request but
rather a single ViewBuilder is created and deals with each request as needed.)

During the re-factor I found an issue with test_servers.py where links were not
containing the project_id like they should. This was fixed and tests were updated.

Overall we had good test coverage so not many tests had to be updated during
the refactor.

(Patch Set 2) Updated _set_request and _get_request to be a request property
              and a set_request method which can be overridden.
(Patch Set 3) Fixed tests that I broke with the last update.
(Patch Set 4) Changed "id in network.keys()" to "id in network"

Change-Id: I034404892018e99987f80789d7f7e406ff31658c