Too much information is returned from POST /servers
Currently we're returning a lot of information in the POST /servers
response. This isn't according to spec and is misleading to other
projects that might try and implement the OpenStack API specification.

The example on docs.openstack.com for the v1.1 API says:

"Note that when creating a server only the server ID, its links, and
the admin password are guaranteed to be returned in the request.
Additional attributes may be retrieved by performing subsequent GETs
on the server."

We're returning too much right now and this patch addresses that.

Further patches should be submitted to refactor the 'view builder'
concept as using keyword arguments like this patch does is not ideal
and I recognize that.

Change-Id: If246e51dbf84f1db3d2905694235692ab027859d